% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_cpm.R
\name{bayes_cpm}
\alias{bayes_cpm}
\title{Bayes CPM model using Stan}
\usage{
bayes_cpm(standata, ...)
}
\arguments{
\item{standata}{a list output from bayesCPM::mkStanDat() containing the data}

\item{...}{other parameters passed to rstan::sampling()}
}
\value{
a 'stanfit' object
}
\description{
get posterior samples from Bayes CPM model using Stan
}
\examples{
dat1_stan <- mkStanDat(dat1, "outcome_var", c("pred1","pred2"), 2)
fit <- bayes_cpm(dat1_stan)
}
