% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayescpm-package.R
\docType{package}
\name{bayescpm-package}
\alias{bayescpm}
\alias{bayescpm-package}
\title{bayescpm: Bayesian approach to clinical prediction model's development and validation}
\description{
BAyesian power, sample size, and value-analysis calcualtor for risk prediction models.
}
\author{
\strong{Maintainer}: Mohsen Sadatsafavi \email{mohsen.sadatsafavi@ubc.ca} (\href{https://orcid.org/0000-0002-0419-7862}{ORCID})

}
\keyword{internal}
