% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPostPred.R
\name{getPostPred}
\alias{getPostPred}
\title{Get posterior predictive values from fit Bayes CPM model}
\usage{
getPostPred(fit, newdata, draws = 5)
}
\arguments{
\item{fit}{output list from bayes_cpm() containing 'stanfit' object and 'standata' data used to fit model}

\item{newdata}{a data frame with columns for each predictor used in the model}

\item{draws}{number of draws to return}
}
\value{
matrix with posterior predictive values
}
\description{
Draw posterior predictive values from a Bayes CPM fit
}
\examples{
fit <- bayes_cpm(y~x1+x2, data=dat)
getPostPred(fit, dat[,c("x1","x2")])
}
