% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getQuantile.R
\name{getQuantile}
\alias{getQuantile}
\title{Get conditional quantile from fit Bayes CPM model}
\usage{
getQuantile(fit, newdata, q = 0.5, summ = TRUE, ...)
}
\arguments{
\item{fit}{output list from bayes_cpm() containing 'stanfit' object and 'standata' data used to fit model}

\item{newdata}{a data frame with columns for each predictor used in the model}

\item{q}{quantile to be estimated (default is median, i.e. q=0.5)}

\item{summ}{logical. Should the function return a summary of the posterior
conditional mean? (default=TRUE)}
}
\value{
posterior conditional quantile summary (summ=TRUE) or values (summ=FALSE)
}
\description{
Calculates the posterior conditional quantile
}
\examples{
fit <- bayes_cpm(y~x1+x2, data=dat)
getQuantile(fit, data.frame(pred1=c(0,1),pred2=c(1,1)), q=0.5)
}
