% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_cpm.R
\name{bayes_cpm}
\alias{bayes_cpm}
\title{Bayes CPM model using Stan}
\usage{
bayes_cpm(
  formula,
  data,
  link,
  dir_prior_conc = function(n) 1/n,
  prior = NULL,
  ...
)
}
\arguments{
\item{formula}{an object of class "formula": a symbolic description of the model to be fitted. outcome should be an ordered factor}

\item{data}{a data frame containing the variables in the model}

\item{link}{the link function to be used. must match one of c("logistic", "probit", "loglog", "cloglog", "cauchit")}

\item{dir_prior_conc}{concentration parameter for Dirichlet distribution as a function of the number of categories, n. The default is 1/n}

\item{prior}{NOT CURRENTLY USED}

\item{...}{other parameters passed to rstan::sampling()}
}
\value{
a list containing a 'stanfit' object and 'standata' used to fit model
}
\description{
get posterior samples from Bayes CPM model using Stan
}
\examples{
fit <- bayes_cpm(y~x1+x2, data=dat1, link="logistic", dir_prior_conc=function(n) 1/n)
}
