% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesCPM-package.R
\docType{package}
\name{bayesCPM-package}
\alias{bayesCPM-package}
\alias{bayesCPM}
\title{The 'bayesCPM' package.}
\description{
Bayesian cumulative probability models (CPMs) for continuous and
   mixed outcomes. This package includes functions to fit a Bayesian CPM using
   the R interface to Stan.
}
\references{
Stan Development Team (2019). RStan: the R interface to Stan. R package version 2.19.2. https://mc-stan.org
}
